// web/src/shared/types/news.ts
export interface Article {
  source: {
    id: string | null;
    name: string;
  };
  author: string | null;
  title: string;
  description: string | null;
  url: string;
  urlToImage: string | null;
  publishedAt: string;
  content: string | null;
}

export interface PagedResponse {
  status: string;
  totalResults: number;
  articles: Article[];
}